Update README.md
Really good tutorial! I enjoyed reading it since it was light in the correct places and serious where it should be. The satisfaction exercise turned the banal wordcount to something entertaining. 

I am not certain about the suggestions in the body of the text but these are the questions that stroke me while reading this.